Added automatic pings to client, robot, and server (500ms delay) to enable automatic reconnection on restart. Reduced display frequency of most messages to once every 500ms to prevent lag. Fixed watchdog timer. Fixed segfault on destruction of Robot object. Fixed missing button value text in DummyRobot::control().

+++ util/watchdog.h
@@ -21,7 +21,7 @@ public:
 
 private:
 	//Variables
-	double interval;
+	unsigned long interval;
 	time_t lastTime;
 };
